+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * Updated React 15.2.1 to version 15.3.1
 * Updated React Native 0.31.0 to version 0.32.0
 * Updated PropTypes declaration to correct array
+* **NEW** DialogReferenceManager
 
 ### 0.6.0 (August 19, 2016)
 * Forked from react-native-popup