+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * Updated React Native 0.31.0 to version 0.32.0
 * Updated PropTypes declaration to correct array
 * **NEW** DialogReferenceManager
+* PropTypes missing from DialogBox have been added
 
 ### 0.6.0 (August 19, 2016)
 * Forked from react-native-popup