+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### HEAD
 
-* Nothing Current.
+* **NEW** DialogReferenceManager
+* PropTypes missing from DialogBox have been added
 
 ### 0.6.4 (February 9, 2017)
 * Updated React peer to 15.4.2
@@ -20,8 +21,6 @@
 * Updated React 15.2.1 to version 15.3.1
 * Updated React Native 0.31.0 to version 0.32.0
 * Updated PropTypes declaration to correct array
-* **NEW** DialogReferenceManager
-* PropTypes missing from DialogBox have been added
 
 ### 0.6.0 (August 19, 2016)
 * Forked from react-native-popup